requirements_new.txt
Flask==3.0.0
python-dotenv==1.0.0
openai==1.3.7
Werkzeug==3.0.1
